move util funcs to separate folder
