+++ app/components/main-content/PrelaunchSignUpForm.tsx
@@ -73,57 +73,6 @@ const PrelaunchSignUpForm: React.FC<PrelaunchSignUpFormProps> = ({ setFlowState 
     setFlowState('confirmed')
   }
 
-const validateMinLength = (minLength: number, field: string) => (value: string) => {
-  return value.length < minLength ? `${field} must have at least ${minLength} letters. Please try again.` : null;
-}
-
-const validateEmail = (value: string) => {
-  return /^\S+@\S+\.\S+$/.test(value) ? null : 'Invalid email format. Please enter a valid email address.';
-}
-
-interface TextInputConfig {
-  field: string;
-  label: string;
-  placeholder: string;
-  required?: boolean;
-  validate?: (value: string) => string | null;
-}
-
-const TEXT_INPUT_CONFIG: TextInputConfig[] = [
-  {
-    field: 'name',
-    label: 'Name',
-    placeholder: 'Enter your name',
-    required: true,
-    validate: validateMinLength(1, 'Name'),
-  },
-  {
-    field: 'email',
-    label: 'Email',
-    placeholder: 'Enter your email',
-    required: true,
-    validate: validateEmail,
-  },
-  {
-    field: 'interests',
-    label: 'Interests',
-    placeholder: 'Enter your interests',
-    validate: validateMinLength(3, 'Interests'),
-  },
-  {
-    field: 'source',
-    label: 'How did you hear about us?',
-    placeholder: 'Enter your source',
-    validate: validateMinLength(3, 'Source'),
-  },
-  {
-    field: 'features',
-    label: 'What features are you most interested in?',
-    placeholder: 'Describe features',
-    validate: validateMinLength(5, 'Features'),
-  }
-];
-
   return (
     <>
       <Box>
